+++ source/extensions/lightspeed.trex.stage_manager.plugin.filter.usd/lightspeed/trex/stage_manager/plugin/filter/usd/extension.py
@@ -22,11 +22,12 @@ from omni.flux.stage_manager.factory import get_instance as _get_factory_instanc
 
 from .is_capture import IsCaptureFilterPlugin as _IsCaptureFilterPlugin
 from .is_category import IsCategoryFilterPlugin as _IsCategoryFilterPlugin
+from .mesh_prims import MeshPrimsFilterPlugin as _MeshPrimsFilterPlugin
 
 
 class LightspeedStageManagerUSDFilterPluginsExtension(omni.ext.IExt):
 
-    _PLUGINS = [_IsCaptureFilterPlugin, _IsCategoryFilterPlugin]
+    _PLUGINS = [_IsCaptureFilterPlugin, _IsCategoryFilterPlugin, _MeshPrimsFilterPlugin]
 
     def on_startup(self, _):
         carb.log_info("[lightspeed.trex.stage_manager.plugin.filter.usd] Startup")