+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.13.0]
+## Added
+- Stage Manager: Added a mesh/instance tab
+
 ## [1.12.1]
 ## Added
 - Stage Manager: Added a categories interaction tab