+++ source/extensions/lightspeed.trex.stage_manager.plugin.tree.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Added
+- Created with stage manager mesh tab
+
 ## [1.0.0]
 ### Added
 - Created