+++ source/extensions/omni.flux.stage_manager.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.5]
+### Fixed
+- Fixed tab padding
+
 ## [1.4.4]
 ### Changed
 - Reduced the tab horizontal padding